golangci-lint fixes for 'system-probe' pkg
