+++ pkg/network/driver_interface.go
@@ -27,9 +27,9 @@ type DriverExpvar string
 
 const (
 	totalFlowStats  DriverExpvar = "driver_total_flow_stats"
-	flowHandleStats              = "driver_flow_handle_stats"
-	flowStats                    = "flows"
-	driverStats                  = "driver"
+	flowHandleStats DriverExpvar = "driver_flow_handle_stats"
+	flowStats       DriverExpvar = "flows"
+	driverStats     DriverExpvar = "driver"
 )
 
 const (
