+++ pkg/network/driver_interface_test.go
@@ -20,10 +20,8 @@ import (
 type TestDriverHandleInfiniteLoop struct {
 	t *testing.T
 	// state variables
-	hasBeenCalled   bool
-	lastReturnBytes uint32
-	lastBufferSize  int
-	lastError       error
+	hasBeenCalled  bool
+	lastBufferSize int
 }
 
 func (tdh *TestDriverHandleInfiniteLoop) ReadFile(p []byte, bytesRead *uint32, ol *windows.Overlapped) error {
@@ -78,17 +76,4 @@ func TestConnectionStatsInfiniteLoop(t *testing.T) {
 		return true
 	})
 	require.NoError(t, err, "Failed to get connection stats")
-}
-
-type TestDriverHandleFiltersSuccess struct {
-	t *testing.T
-	// state variables
-	hasBeenCalled   bool
-	lastReturnBytes uint32
-	lastBufferSize  int
-	lastError       error
-}
-
-func (tdh *TestDriverHandleFiltersSuccess) ReadFile(p []byte, bytesRead *uint32, ol *windows.Overlapped) error {
-	return nil
 }
