+++ pkg/network/tracer/tracer_windows_test.go
@@ -14,10 +14,6 @@ import (
 	"testing"
 )
 
-func dnsSupported(t *testing.T) bool {
-	return true
-}
-
 func httpSupported(t *testing.T) bool {
 	return false
 }
